Theme: do not set background of 'pre' elements
While the result looks good with the dark theme, it makes text
unreadable with the light theme. A workaround of this is setting the
'color' as well, but that would only set the main text color, leaving
other low-contrast colors.

This is not a final solution: 'pre' elements need to be fully restyled.